@@ -10,10 +10,9 @@
         }
     ],
     "require": {
-        "php": "^7.2|^8.0",
+        "php": "^8.1",
         "ext-openssl": "*",
-        "ext-json": "*",
-        "jms/serializer": "^3.0"
+        "ext-json": "*"
     },
     "autoload": {
         "psr-0": {
