@@ -2,4 +2,8 @@
 
 namespace AdAuth\Response;
 
-class PingResponse extends AbstractResponse { }
+class PingResponse extends AbstractResponse {
+    public function __construct() {
+        parent::__construct(true);
+    }
+}
